refactor(server): group connector grant-type policy with connectorauth
ConnectorGrantTypes, GrantTypeAllowed and parseScopes are dex connector policy
and input helpers, not OAuth2 protocol vocabulary, so they don't belong in the
oauth2 package. Move them out of the oauth2.go/server.go grab bags into
connectorauth.go, whose stated purpose is connector authorization and which
already documented GrantTypeAllowed. No API change; all three stay in package
server.
